backend/pyproject.toml
[tool.poetry]
name = "polyglot-backend"
version = "0.1.0"
description = "Backend for Polyglot Universal Code Translator"
authors = ["Polyglot Team"]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-cors = "^4.0.0"
google-genai = "^0.2.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
